pull env-specific tests from CI/CD and run locally only
